pbs api types: relax influx organization and bucket length
Relax the length limit for organization and bucket name for InfluxDB
metric servers in Proxmox Backup Server.

Commit 57fa20406 ("pbs-api-types: add metrics api types") introduced
the api schema definition for InfluxDB metric servers, limiting the
name of the organization and the name of the bucket to a length of 3,
most likely as a result of reusing the same values as for the
corresponding config id schema restrictions.

This is however not enforced by the InfluxDB REST api [0, 1] and
stricter than what is defined in Proxmox VE [2].

Reported in the community forum [3].

[0] https://docs.influxdata.com/influxdb/v2/api/#operation/PostOrgs
[1] https://docs.influxdata.com/influxdb/v2/api/#operation/PostBuckets
[2] https://git.proxmox.com/?p=pve-manager.git;a=blob;f=PVE/Status/InfluxDB.pm;h=13a96711e766e2f1ea71424ed6d9d8ec8450504c;hb=HEAD#l24
[3] https://forum.proxmox.com/threads/162521/

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>