generator: fix methods with url params but no query string
Commit ba89c784 introduced a regression where endpoints that have URL
parameters ('/url/with/{param}') would not get generated with format!
calls. This causes URL endpoints such as get_task_status to fail,
since those parameters do not get substituted.

Add a helper for printing the URL that handles all cases and use the
helper instead of generating the URL directly in the respective
methods.

Fixes: ba89c784 ("generator: avoid useless format!() calls")
Suggested-by: Wolfgang Bumiller <w.bumiller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>